sada/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-modeltranslation==0.18.11
Pillow==10.0.1
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
